+++ src/main/java/com/revolvingmadness/sculk/backend/Colors.java
@@ -1,0 +1,8 @@
+package com.revolvingmadness.sculk.backend;
+
+public class Colors {
+    public static final int ERROR = 16007990;
+    public static final int INFO = 9489145;
+    public static final int INTERNAL_ERROR = 13840175;
+    public static final int WARN = 16754470;
+}